requirements.txt
pandas>=1.5.0
numpy>=1.21.0
yfinance>=0.2.0
requests>=2.28.0
python-dotenv>=0.19.0
scipy>=1.9.0
matplotlib>=3.5.0
seaborn>=0.11.0
